+++ pkg/osinfo/osinfo_linux.go
@@ -1,4 +1,4 @@
-package osstats
+package osinfo
 
 import (
 	"bytes"
@@ -7,15 +7,15 @@ import (
 	"time"
 )
 
-func GetOSStats() (OSStats, error) {
-	out, err := _getStats()
+func GetOSInfo() (OSInfo, error) {
+	out, err := getInfo()
 	for strings.Contains(out, brokenPipeOutput) {
-		out, err = _getStats()
+		out, err = getInfo()
 		time.Sleep(retryWaitTime)
 	}
 	osStr := replaceLineTerminations(out)
 	osInfo := strings.Split(osStr, " ")
-	oss := OSStats{
+	oss := OSInfo{
 		Version:  osInfo[1],
 		Platform: osInfo[2],
 		OS:       osInfo[3],
@@ -23,7 +23,7 @@ func GetOSStats() (OSStats, error) {
 	return oss, err
 }
 
-func _getStats() (string, error) {
+func getInfo() (string, error) {
 	cmd := exec.Command("uname", "-srio")
 	cmd.Stdin = strings.NewReader("some input")
 	var out bytes.Buffer
